Mude las clases de Actas a Actas nuevamente. Empeze a modificar las actas de modo que se puedan cargar mas de un bombero a cada tipo de acta.
