+++ actas/choices.py
@@ -1,0 +1,10 @@
+# -*- coding: UTF-8 -*-
+from django.utils.translation import ugettext_lazy as _
+
+
+TIPO_SANCION = (
+    ('advertencia', _("Advertencia")),
+    ('suspencion', _("Suspención")),
+    ('baja', _("Baja")),
+    ('exoneracion', _("Exoneración")),
+)
