+++ bomberos/choices.py
@@ -2,13 +2,6 @@
 from django.utils.translation import ugettext_lazy as _
 
 
-TIPO_SANCION = (
-    ('advertencia', _("Advertencia")),
-    ('suspencion', _("Suspención")),
-    ('baja', _("Baja")),
-    ('exoneracion', _("Exoneración")),
-)
-
 GENERO = (
     ('M', _('Masculino')),
     ('F', _('Femenino')),
